requirements.txt
streamlit
openai-whisper
torch
numpy==1.26.4
scipy
ffmpeg-python
streamlit-webrtc
soundfile
yt-dlp
